+++ @ornikar/jest-config-react-native/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.4.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config-react-native@8.3.3...@ornikar/jest-config-react-native@8.4.0) (2023-09-07)
+
+
+### Features
+
+* **deps:** update dependency jest-expo to v49 ([#963](https://github.com/ornikar/shared-configs/issues/963)) ([299390f](https://github.com/ornikar/shared-configs/commit/299390fd96610148a7ee9d5dc400c81a4d759b25))
+
+
+
+## [8.3.3](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config-react-native@8.3.2...@ornikar/jest-config-react-native@8.3.3) (2023-08-28)
+
+**Note:** Version bump only for package @ornikar/jest-config-react-native
+
+
+
+
+
+## [8.3.2](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config-react-native@8.3.1...@ornikar/jest-config-react-native@8.3.2) (2023-08-01)
+
+**Note:** Version bump only for package @ornikar/jest-config-react-native
+
+
+
+
+
 ## [8.3.1](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config-react-native@8.3.0...@ornikar/jest-config-react-native@8.3.1) (2023-07-26)