+++ @ornikar/repo-config-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.1.5](https://github.com/ornikar/shared-configs/compare/@ornikar/repo-config-react@10.1.4...@ornikar/repo-config-react@10.1.5) (2023-09-13)
+
+**Note:** Version bump only for package @ornikar/repo-config-react
+
+
+
+
+
+## [10.1.4](https://github.com/ornikar/shared-configs/compare/@ornikar/repo-config-react@10.1.3...@ornikar/repo-config-react@10.1.4) (2023-08-28)
+
+**Note:** Version bump only for package @ornikar/repo-config-react
+
+
+
+
+
 ## [10.1.3](https://github.com/ornikar/shared-configs/compare/@ornikar/repo-config-react@10.1.2...@ornikar/repo-config-react@10.1.3) (2023-07-25)
 
 **Note:** Version bump only for package @ornikar/repo-config-react